[DTLTO] DTLTO.cpp had been splitted into three sources: DTLTOInputFiles.cpp, DTLTO.cpp and DTLTODistributionDriver.cpp.
