+++ llvm/lib/DTLTO/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_llvm_component_library(LLVMDTLTO
+  DTLTOInputFiles.cpp
   DTLTO.cpp
+  DTLTODistributionDriver.cpp
 
   LINK_COMPONENTS
   BinaryFormat
